refactor: put functions in the right files
