+++ src/app/shared/helpers/functions.helper.ts
@@ -1,3 +1,5 @@
+import { formatDate } from '@angular/common';
+
 export const getFirstTwoInitials = (name: string) => {
   return name
     .split(' ')
@@ -16,7 +18,7 @@ export const sortByKey = <T extends Record<string, any>>(
   columnKey: keyof T,
   sortOrder: 'asc' | 'desc'
 ): void => {
-  data.sort((a, b) => {
+  data?.sort((a, b) => {
     const valueA = String(a[columnKey]).toLowerCase();
     const valueB = String(b[columnKey]).toLowerCase();
 
@@ -28,22 +30,20 @@ export const sortByKey = <T extends Record<string, any>>(
   });
 };
 
-const routeValueKey: { [key: string]: string } = {
-  dashboard: 'Dashboard',
-  lecturers: 'Lecturers',
-  students: 'Students',
-  internships: 'Internships',
-  'access-control': 'Access Control',
-  courses: 'Courses',
-  Location: 'Location',
-  Annoucements: 'Annoucements',
+export const formatDateToDDMMYYYY = (date: string): string => {
+  const parsedDate = new Date(date);
+  if (isNaN(parsedDate.getTime())) return date; // Return original if invalid date
+  return formatDate(parsedDate, 'dd/MM/yyyy', 'en-US'); // Format date
 };
 
-export function setPageHeader(currentRoute: string): string {
-  for (const key in routeValueKey) {
-    if (currentRoute.includes(key)) {
-      return routeValueKey[key];
-    }
-  }
-  return '';
+export function searchArray<T>(
+  array: T[],
+  searchTerm: string,
+  keys: (keyof T)[]
+): T[] {
+  return array?.filter((item) =>
+    keys?.some((key) =>
+      item[key]?.toString()?.toLowerCase()?.includes(searchTerm?.toLowerCase())
+    )
+  );
 }
